Major refactoring of mbtiles internals and tile format detection
